Extract signing stuff out of clock
It has no need to be on the clock object, and looks a little out of place in the clock.go file.
