+++ internal/merkle/clock/clock.go
@@ -24,8 +24,6 @@ import (
 	"github.com/sourcenetwork/corelog"
 	"github.com/sourcenetwork/immutable"
 
-	"github.com/sourcenetwork/defradb/acp/identity"
-	"github.com/sourcenetwork/defradb/crypto"
 	"github.com/sourcenetwork/defradb/datastore"
 	"github.com/sourcenetwork/defradb/internal/core"
 	coreblock "github.com/sourcenetwork/defradb/internal/core/block"
@@ -113,7 +111,7 @@ func (mc *MerkleClock) AddDelta(
 	}
 
 	if EnabledSigningFromContext(ctx) {
-		err = mc.signBlock(ctx, dagBlock)
+		err = signBlock(ctx, mc.blockstore, dagBlock)
 		if err != nil {
 			return cidlink.Link{}, nil, err
 		}
@@ -218,60 +216,6 @@ func encryptBlock(
 	return &coreblock.Block{Delta: clonedCRDT, Heads: block.Heads, Links: block.Links}, nil
 }
 
-func (mc *MerkleClock) signBlock(
-	ctx context.Context,
-	block *coreblock.Block,
-) error {
-	// We sign only the first field blocks just to add entropy and prevent any collisions.
-	// The integrity of the field data is guaranteed by signatures of the parent composite blocks.
-	if block.Delta.IsField() && block.Delta.GetPriority() > 1 {
-		return nil
-	}
-
-	ident := identity.FromContext(ctx)
-	if !ident.HasValue() {
-		return nil
-	}
-
-	blockBytes, err := block.Marshal()
-	if err != nil {
-		return err
-	}
-
-	var sigType string
-
-	switch ident.Value().PrivateKey.Type() {
-	case crypto.KeyTypeSecp256k1:
-		sigType = coreblock.SignatureTypeECDSA256K
-	case crypto.KeyTypeEd25519:
-		sigType = coreblock.SignatureTypeEd25519
-	default:
-		return NewErrUnsupportedKeyForSigning(ident.Value().PrivateKey.Type())
-	}
-
-	sigBytes, err := ident.Value().PrivateKey.Sign(blockBytes)
-	if err != nil {
-		return err
-	}
-
-	sig := &coreblock.Signature{
-		Header: coreblock.SignatureHeader{
-			Type:     sigType,
-			Identity: []byte(ident.Value().PublicKey.String()),
-		},
-		Value: sigBytes,
-	}
-
-	sigBlockLink, err := putBlock(ctx, mc.blockstore, sig)
-	if err != nil {
-		return err
-	}
-
-	block.Signature = &sigBlockLink
-
-	return nil
-}
-
 // ProcessBlock merges the delta CRDT and updates the state accordingly.
 func (mc *MerkleClock) ProcessBlock(
 	ctx context.Context,
@@ -341,20 +285,4 @@ func (mc *MerkleClock) updateHeads(
 	}
 
 	return nil
-}
-
-type enabledSigningContextKey struct{}
-
-// ContextWithEnabledSigning returns a context with block signing enabled.
-func ContextWithEnabledSigning(ctx context.Context) context.Context {
-	return context.WithValue(ctx, enabledSigningContextKey{}, true)
-}
-
-// EnabledSigningFromContext returns true if block signing is enabled in the context.
-func EnabledSigningFromContext(ctx context.Context) bool {
-	val := ctx.Value(enabledSigningContextKey{})
-	if val == nil {
-		return false
-	}
-	return val.(bool)
 }
